feat: 可运行的history reuse + environment light sampling; refactor: 这就是C艹;

+++ applications/3d_viewer/3d_viewer.cpp
@@ -565,7 +565,7 @@ void Start (std::unique_ptr<MIInfraInterface> && infra, const MainLoopStartConfi
                         }
                         ImGui::Unindent(20);
                     };
-                    DrawTree(0, time_periods.size(), 0);
+                    DrawTree(0, (int)time_periods.size(), 0);
                 }
                 ImGui::End();
             }
